requirements.txt
openai>=1.0
anthropic>=0.30
pyyaml
httpx
pytest
pytest-asyncio
